HPCC-13168 Revert default compression for fixed length to RDiff
In all OSS versions, LZW has been the unintentional default for
compressed disk writes. In the legacy system, if the output was
fixed length, row diff compression was used.
This delivers signiciantly faster compression at the expense of
a lower compression ratio.

This commit reverts the originally intended defaults of RDiff
for fixed length and LZW for variable.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>