HPCC-13205 Avoid spilling window in LOCAL LOOKUP JOIN
(This is seperate PR from PR for 5.0.8 - to avoid merge clash)

There was a window, between successfully collecting the local
rows and allocating the HT, where low memory could cause the RHS
to spill on a local lookup join and attempt a std. local join.
It would then crash unless the generated helper contained a
queryCompareLeft() operator.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>