HPCC-13390 Return persist/superowners in WsDfu.DFUQuery
Both persist and superowners are added into the file
information for ESP Browse Logical Files. ECLWatch
UI may need to display those two columns. 1. The option of
returning superowners is added into WsDdu.DFUQuery request.
If the option is set to true, DFUQuery retrieves the
information about superowners. 2. A SerializeFileAttrOptions
structure is added to dadfs. For now, the includeSuperOwners
is only member of the structure. More options may be added
as needed. The structure is used by serializeFileAttributes()
to decide whether the superowners should be serialized or not.
3. The CFileScanFilterContainer class is enhanced to read
the SerializeFileAttrOptions from IterateFiles filters. The
class is renamed as CIterateFileFilterContainer since it
not only has the filters for FileScan and also has the
SerializeFileAttrOptions. 4. The call to CIterateFileFilter
Container.readFilters() is moved out of the scan() since the
SerializeFileAttrOptions is used by the serializeFileAttributes().

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>