HPCC-13134 Record fields with complex xpaths should output valid xml
This commit will cause scalar record fields with complex xpaths to
use the field name rather than xpath to output xml and json.

The caveat being that the record still can't be used to read the output
back in, but there is no way to guess the original format from the xpath
so it's better not to try to output complex xml when its quite likely
to be wrong.  In any case the output should be valid xml.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>