HPCC-13440 Backupnode could corrupt slaves file and hit problems
If Dali is down and coming up when backupnode is started, the
daliadmin call that the backupnode script makes issues warnings
about failing to connect to Dali, which are inadvertently
captured into the slaves file.
When Dali comes up, the slaves file is then corrupted.
Backupnode continues to use the slaves file, not aware that is
invalid entries and hits problems.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>