HPCC-13439 Query of LDAP maxPwdAge issued too often
Currently the call to query the LDAP password expiration "maxPwdAge" is a
member of the CLdapClient class, which is instantiated numerous times.
Although cached, this value rarely (or never) changes, so there are too
many calls to LDAP to query it. This pull request implements a shared
accessor call that all instances of CLdapClient can call to query it.

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>