HPCC-13470 memory leaks in ldapsecurity code
CLdapSecManager::createResourceList allocates resources that are stored
by the CAuthMap object in an IArrayOf<CSecResourceListHolder> m_resourcelists
object. When the CAuthMap gets destroyed, and its m_resourcelists array, the
contents of that array are not released and therefore leak. This pull request
adds a virtual destructor to CAuthMap that iterates over the list and releases
each one.

The other leak referenced in the Jira bug was confirmed not to be a leak. To
prove, I added PROGLOG messages to the CTOR and DTOR of CLdapSecUser, and
ensured that the CLdapSecResource::clone() was called (hundreds of times).
Ran this test many times/variations, and in all cases the objects were
correctly destroyed


Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>