HPCC-13483 Remove call to exportWorkUnitToXML() in WUInfo
For a large WU, the call will be expensive. By this fix,
WsWorkunits.WUInfo will not call exportWorkUnitToXML() for
returning WU XML size. Legacy ECLWatch will always show
the option for the XML download. If a user elects to
display the XML and the size of the XMl is > 5M bytes, the
display function returns an error message about large XML
size.

BTW, the existing new eclwatch always has the option for
the XML download.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>