HPCC-13495 Use correct default target directory for despray
The existing ESP code calls the RemoteFilename::setPath()
to set the target directory for despray. If a user does
not specify the target directory, the setPath() uses the
/var/lib/HPCCSystems/myesp as default target directory. In
this fix, If a user does not specify the target directory,
I use the target IP to retrieve the drop zone directory on
that IP and use the drop zone directory to be default
target directory.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>