HPCC-13532 Set reschedule flag based on WU state
The EventSchedule flag in ECL WUInfo response is used to
decide whether a Reschedule button should be shown on the
ECL WU Details page or not. Since the schedule is not
allowed for WUStateFailed, WUStateArchived, WUStateAborting,
WUStateAborted, and WUStateScheduled, the EventSchedule flag
should be set based on WU state.

This fix also removes the duplicated code in getWorkflow().

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>