HPCC-13545 Expose lock info to clients.
Previously locking information was available only en-bulk via
a single output processed by Dali and returned as a formatted
string table.
There was a couple of places (in daliadmin and dalidiag) that
worked hard to parse this formatted string, to perform
filtering.

This change allows path and ip filtering of locks to be
handled at the server, for the result to be serialized to the
requesting client and then deserialized into interrogatable
interfaces.
This will also allow clients (e.g. Eclwatch) to request and
format the locking information as they see fit.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>