HPCC-13545 Separate exposing data and formating in getLocks
The existing getLocks() calls the getLockInfo() to retrieve lock
data and create a formated string on dali server side. In this
fix, the getLockInfo() is divided into two parts, one for
retrieving lock data and one for creating a formated string. A
new getLocks() is created to call the part for retrieving lock
data only. Both parts are built into a new lock data helper class.
A new getLocks() is created to use the helper class to retrieve
serialized lock data without formating. After the data is
received, the client sides (daliadmin and dalidiag) uses the
helper class to create formatted strings as needed. The
getLockInfo() is still used by dali server for other functions
(such as log),

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>