HPCC-13608 Return error message if too many files
In some environment with many logical files, the "Jbuf: out of
memory" occurs when sorting the files on ESP for "Browse Logical
Files". As a short term fix, instead of returning all files, dali
will return a random N files with total number of all files to
client (ESP). (dali client may specify the N value and a staring
point of the N). As a dali client, ESP forwards the N files (with
pages) to its client with a warning message of "More than N files
found..." if total number of files > N.

A long term fix will be done in HPCC-13425 later.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>