HPCC-12914 Change esp log settings when ESP is running
By this fix, an admin user may change ESP log settings for a
running ESP through a new ESP service: WSESPControl. For
example, the user may change ESP log level. If the
controlPort is not set in environment.xml, the espcontrol
should be loaded and on port 8010. Turn espcontrol off only
if the admin sets controlPort == 0.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>