HPCC-13496 Fix compilation Error: 'this' pointer and references cannot be null in well-defined C++ code
The this and references are checked for null to protect against unexpected behaviour.  However, this and references cannot be null in well-defined code and some compilers such as Clang will refuse to compile such code (unless the compiler behaviour is overrriden). The code may also be optimized away eliminating the expected protection

New function isNull() created to check this and references for null.  All direct comparisons with null replaced with a call to this function.