HPCC-13642 Reduce out of memory errors when heaps are highly fragmented
When a fixed size heap does not contain any more space it attempts to
allocate a new page.  If no pages are available the heap manager calls
any registered callbacks to free memory and then tries again to allocate
a new page.

Unfortunately, if a callback frees many rows, but no complete pages, then
it does not allow the allocation of a new page to succeed.  (Row distributions
like this are relatively common for send rows buffered for a distribute.)

This patch adds an attempt to allocate a row from the heap if no page was
available (providing some rows were freed from this heap when the callback was
called).

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>