HPCC-13618 Add a useful exception toaster message
If a cluster does not exist when choosing an alternative 'allowed cluster' within workunit details the toaster message that appears does not provide clear information as to what the exception is. In this case, it returns "GMT: Invalid cluster name: ClusterName" which is helpful. Improving of exception handling may need to be done in other cases throughout the application.

Signed-off-by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>