HPCC-12224 eclagent.log sometimes shows activity not related to workunit
Currently the code that extracts the eclagent logfile entries from the master
eclagent logfile keys off of the process ID. However it scans the entire
logfile line looking for that PID, so if it happens to match the TID or any
other random string it will include that line in the output. This PR determines
the offset of the PID entry and makes sure to only include lines that match
the PID at that offset.

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>