HPCC-13608 Return warning message if too many files
In some environment with many logical files, the "Jbuf: out of
memory" occurs when sorting the files on ESP for "Browse Logical
Files". As a short term fix, instead of returning all files, dali
will return N files with total number of all files to client (ESP
). (dali client may specify the N value). As a dali client, ESP
forwards the N files (with paging) to its client with a warning
message of "More than N files found..." if total number of files
> N.

Changes after review: (1) Set max number (N) of files to return,
not set from/to. Remove the new cache which is no longer needed.
(2) Move IElementsPagerEx::getTotalElements() into IElementsPager;
(3) the 'total' is adjusted by the postfilter; (4) implement the
getTotalElements() for WUs/Queries.

A long term fix will be done in HPCC-13425 later.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>