HPCC-13591 Tell user abort active WU before deleting
When a user wants to delete an active WU, the existing code
sets the WU state to failed and, then, deletes it. That may
get other weird behaviour like running workunits unable to
load their DLLs, etc. This fix removes the code for setting
the WU state to failed and tells the user to abort the
active WU before deleting.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>