HPCC-13692 Not considering globalness of loop graph correctly
Checking whether a graph required global synchronization or not,
was performed too early in the case of LOOP graphs, before the
graph was fully initialized. As a result, it did not walk the
graph and correctly decide if the loop body was global or not.
Consequently in some cases, where no other dependency or activity
made the graph global, the job would hang.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>