HPCC-13608 Return warning message if too many files
In some environment with many logical files, the "Jbuf: out of
memory" occurs when sorting the files on ESP for "Browse Logical
Files". As a short term fix, instead of returning all files, dali
will return to client (ESP) N files and a flag which indicates
that the N files are a subset of all matches. (The N value may be
specified by a client). ESP forwards the N files (with paging) to
its client with a warning message about the subset of files.

A long term fix will be done in HPCC-13425 later.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>