HPCC-13791 Use keys in sort partitioning
Previously the sort partitioning used in global sort/join was
serializing whole rows to and from the slaves/master, instead of
just the key fields it needed. This was costly in terms of
transmission time, but also required an excessive amount of
memory where rows were big, ultimately limiting the size that
sort could cope with before running out of memory on the master
whilst holding all the partition points.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>