HPCC-13796 Add support for filtering WUs by file in Cassandra
Store the fileread information in a child table, and create an index that can
be used to lookup wuids from filenames.

When filtering by filename, use the latter table to generate a list of wuids
and then merge and fetch them.

Also fixed some issues in the behaviour of the paging code and the test
program.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>