HPCC-13805 Can't use parameter in embed option string
Various issues with previous attempt to fix this, notably that the parameter
in question is still passed to the embedded code where it is unlikely to be
expected.

Users can achieve te desired result using a FUNCTION wrapper around the embed,
so issue an improved error message indicating that a FUNCTION can be used.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>