HPCC-13899 Add support for graph progress in Cassandra
Refactor workunit graph progress API to make it easier to separate out
code that needs to be back-end dependent, and also make the existing
implementation a bit cleaner, easier to use, and more efficient.

Add Cassandra implementation.

Add unit tests.

Fix crash in configgen when Cassandra workunit support is requested.

Clarify intent of code that was relying on a ptree bug.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>