HPCC-13883 ulUnicodeLocaleEditDistanceWithinRadius can access uninitialized data
ulUnicodeLocaleEditDistanceWithinRadius implements an optimized version of
LevenshteinDistance algorithm which works correctly only when difference in
the number of characters of strings being compared does not exceed the radius.
Current implementation of ulUnicodeLocaleEditDistanceWithinRadius does not
enforce that restriction when Locale is provided, and enforces it incorrectly
when Locale is not provided.

Add code to correctly enforce the above restriction for
ulUnicodeLocaleEditDistanceWithinRadius with and without Locale.


Signed-off-by: Edin Muharemagic <edin.muharemagic@lexisnexis.com>