HPCC-13977 Fix issues running thor/hthor regression suites using C* workunits
Fix various issues related to setting graph states and timing of setting
results that were revealed when running the regression suite using Cassandra
workunit support.

It's arguable whether it should be permissible to set a result after releasing
the workunit of the result - should the result link the workunit, for example,
and thus prolong the lock that the workunit holds until the result is
released. However at present it only worked in Dali by luck (and the lock
would be released before the result was written) and not at all in Cassandra,
while the minor refactoring of the thor/hthor code SHOULD result in more
efficient code as well as being clearer (and working) so I am ignoring the
philosophical question for now.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>