HPCC-13731 WorkunitsService does not always go via workunit.hpp
I suspect it's going straight to the XML to try to common up between dali and
sasha workunits, but the code that pulled the wu XML from Sasha was broken so
that feature never worked anyway. In the interests of
portability/maintainability I propose to remove the (broken) sasha capability.
With a bit of luck once Cassandra is in widespread use for workunits we won't
need to archive them any more anyway.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>