HPCC-13984 Increase listen backlog for large clusters with many thor slaves
Added check for kernel somaxconn and env mpSoMaxConn at MP start up.
Added changes to conf and xml file for mpStart, mpEnd, mpSoMaxConn to
not be nested so that configmgr will use with new configs.
Fixed typo in xml.in.
Updated testing/regress/environment.xml.in to match same changes.

Signed-off-by: Mark Kelly <mark.kelly@lexisnexis.com>