HPCC-14005 WuTest selftest is not deleting one of the wus
Added code to set the state to completed so that the WU can be deleted - this
revealed another issue in that the program blocks on exist waiting for a
response from 1.2.3.4 where it is trying to delete a file.

Added code to remove the file info from the query before deleting - this
revealed another issue where information could only be added to child tables
in Cassandra and was never deleted.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>