HPCC-13958 Return CanReplicate in WsDfu.DFUInfo
This fix checks and returns whether a logical file can
be replicated on a cluster or not. The information will
be used by new ECLWatch to show a Replicate button on
the Logical File Details page. The file can only be
replicated on a cluster if: 1. the cluster is a thor
cluster; 2. the cluster has multiple nodes; and 3. the
@replicateOutputs attribute is set to true.

Change based on review: 1. Cache node group info to a
hash table. 2. the default timeout value of the cache
can be modified using ConfigMgr.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>