HPCC-14067 Avoid possible deadlock if spilling + callback called
When the memory manager is actively trying to free memory at the
same time that a spilling stream is spilling and in the process
of deregistering it's callback a deadlock can occur - it can
block on a rows lock.
Avoid by checking if already spilt upfront in the callback and
exiting.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>