HPCC-14063 Nagios Toolbar icon out of sync
Whenever ECL Watch was fetching the latest nagios status if a state change was present the toolbar icon would not update. This was due to removing a class that was possibly not present. Therefore, we will remove all classes and set the class based on the response.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>