HPCC-14093 Fix regression causing corrupt dafilesrv commands
The throttling refactoring (see HPCC-13573), introduced a couple
of issues, which caused commands to be corrupted.
a) The 1st part(s) of partial messages, handled by select
notifier were lost, with only the last fragment being submitted
as a command. Only really noticable for large messages.
b) There was a transfer of MemoryBuffer content using swapWith,
which caused a buffer to subsequently used with the wrong
endianess.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>