HPCC-14095 Fix race condition on shared spilling stream.
There was a small window, where a shared stream is requested
and it has yet not spilled, if during the creation of the stream
it spills, the stream will not have registered its callback and
will not have been told it's relative position in the spill file.
Then when the stream is read an assertion is hit.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>