HPCC-12687 Roxie may retry indefinitely in localSlave mode
If localSlave mode is selected and multiple channels are in use, then (a)
any request for a channel other than 1 is likely to result in an
"Unregistered query" exception, and (b) the message is then requeued for
retry and an infinite loop results.

There is no point sending retries to 'buddies' in localSlave mode since there
are no buddies. Also, put in a check at startup that localSlave is not used
with multiple channels.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>