HPCC-14115 LDAP Enumerations need to support paging
LDAP imposes a limit of 1,000 on the number of entries that can be returned
from any search. Once there are over 1,000 matches to any search, an error
is returned and you cannot even get the first 1000.  This PR implements a
CPagedLDAPSearch class, which wraps code to perform an LDAP Paged Search,
and uses it in any ILDAPClient calls that could return this many entries.
It also implements a countEntries method, which is used in several places
within LDAP clients.  This includes:

retrieveUsers
getResources
getResourcesEx
getAllGroups
getGroupMembers
countEntries

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>