HPCC-12872 TRACE activity
Implement a new TRACE activity.

Syntax is:

myds := TRACE(ds [, traceOptions]);

The available options are:

 - Zero or more expressions, which act as a filter. Only rows matching the
   filter will be included in the tracing.
 - KEEP (n) indicating how many rows will be traced.
 - SKIP (n) indicating that n rows will be skipped before tracing starts.
 - SAMPLE (n) indicating that only every nth row is traced.
 - NAMED(string) providing the name for the rows in the tracing.

Tracing is output to the log file, in the form
 TRACE: <name><fieldname>value</fieldname>...</name>

Tracing is not output by default even if TRACE statements are present - only
if the workunit debug value traceEnabled is set or if the default platform
settings are changed to always output tracing. In Roxie you can also request
tracing on a deployed query by specifying traceEnabled=1 in the query XML. It
is therefore possible to leave trace statements in the ECL without any
detectable overhead until tracing is enabled.

It is also possible to override the default value for KEEP at a global,
per-workunit, or per-query level.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>