HPCC-14155 Avoid slaves unregistering after clean shutdown request
The Thor master tells the slaves to shutdown, but doesn't wait for
acknowledgement and quits, the thor scripts then send a SIGTERM
signal to the slaves, which (in a race condition) can cause the
slaves to attempt to unregister from the master, by which time
the master has quit.
As a result the slaves throw an unregistration error and log it.

This change waits for the slaves to acknowledge the shutdown and
flags not to unregister if a signal is subsequently received.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>