HPCC-14212 Avoid deadlock with subscription sub system
Deleting a subscription stub can involve keeping it's mutex and
calling back to remove NodeSubscriber implementation requiring
the NodeSubscriber mutex to be locked.
Conversely, removing the NodeSubscriber involves having the
NodeSubscriber mutex locked and can cause the subscription sub
system to require it's lock too. If 2 threads are in each of
these states, deadlock can ensue. It's a race condition, if the
notification mechanism is still actively notifing it will have
the subscriber linked and not hit this issue, it will end up
destroying the subscriber outside of the mutex deadlock scenario.

Fix by passing ownership of the subscriber objects to the
notification mechanism.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>