HPCC-14448 LDAP Creation of parent OU should be more lenient
If the LDAP Admin account does not have create access to any node in a given
OU path, then it throws and terminates. Typically, these admin accounts will
be created to only have create access to a subbranch of the OU tree, and the
parent branches are created via iSIT tickets. This PR checks for the existence
of a given OU before trying to create it, eliminating the exception.

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>