HPCC-14468 LDAP SecMgr unable to search distinguishedName with backslash
Using OpenLDAP, if a DistinguishedName has a backslash escape character in it such as
CN=Whitehead\, William,OU=BCT,OU=Users
then calls to ldap_search_ext_s fail with a "bad filter" return code. According
to LDAP RFC 4515, these characters should be encoded as \xx , where xx is the
hex ascii code for that character. These characters are \,(,),and *

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>