HPCC-14367 Support target alias names for roxies behind a VIP
Roxies accessed from behind a vip usually use seperate target names to
represent the same set of queries.  Each roxie can also run more than
one target at the same time.  A way is needed to specify that any of the
targets representing the same set of queries can be used.

Allowing targets that represent the same set of queries to have a shared
alias will allow callers to select the right set of queries when calling
through a VIP.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>