HPCC-14584 Avoid potential exception in call from destructor
With the c++11 semantics enabled, throwing an exception directly
or indirectly from a desctructor that hasn't been declared with
noexcept(false) will terminate the program.

The locking code prior to this change had a wrapper class that
during destruction called a method that could throw an exception.
Avoid the use of this simple wrapper class and call the methods
it was calling in the ctor and dtor directly to avoid being hit
by the new c++11 semantics.
NB: alternative fix was to add noexcept(false) to dtor, but
avoiding the possibility of an exception in destructor altogether
is safer.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>