HPCC-14730 Fix issue with conditional null inputs in child queries
If a child query has an activity whose input could be evaluated
to null (an example is Filter with !canMatchAny()), if that
activity calculates that it has a null input on the 1st execution
but not on subsequent exections, Thor hits a runtime assertion
whilst connecting the graph, specifically: "assert(outLink)"

This was because once the input was deemed as null, the acitivty
wasn't fully initialized. When subsequent executions of the child
query relied on the activitiy, it was still not initialized and
consequently had not added it's output to be connected in the
graph.

Fix is to treat the activity normally during initialization and
special case the null activity connection during connect only.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>