HPCC-14515 Issues sorting logical files
There was previous issues sorting within the SearchResultsWidget we issues a fix for that which in turn caused issues in our DFUQueryWidget. Modified the service call to accomodate both areas. Also, "null" is returned at times for certain files which causes sorting issues. Will assign to 0 on a per row basis but not display 0.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>