HPCC-14500 Change warning message for cluster status
To get the status information for a cluster, WsSMCEx::Activity is
using the queue name of the cluster to looks for the cluster in the
"/Status/Servers" section of DALI. If it cannot be found, WsSMCEx::
Activity returns a warning: "cluster not attached". This fix changes
the warning to: "Not listening for workunits".

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>