HPCC-14858 Give context to OOM if upstream from gathering activity
If a upstream activity requires a group or set of rows to be in
memory and cannot spill. Give some context to a OOM exception
that it catches.

This also changes the way the original encountered exception is
reported. Before this commit, the activity that hit the exception
would fire the exception to the master asap and abort the query.
To catch and add context, that had to change.
Now the exception is propagated up the activity chain (was before)
before firing the exception back to the master.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>