HPCC-14164 Error: System error: 1000: Key size mismatch on key [merger]
Not been able to reproduce the error, but it's clearly from the symptoms
related to keyedSIze being uninitialized - this could happen if one of the
keys being merged had zero parts - for example an empty index.

Initializing the variable properly and avoiding the check that failed in the
uninitialized case should help (or at least reveal someother bug).

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>