HPCC-14306 ECL Watch remembers checked files
Whenever a user is using the filtering capabilities of dgrid the files previously checked are still checked. If a user decides to delete a file before the filter was set two files will now be checked. Adding an additional fail-safe that clears selection on refresh.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>