HPCC-14165 LDAP compare on 389 DirectoryServer fails
Using 389DirectoryServer, when determining whether or not a user is a member
of a group (user/MemberOf), the security manager calls ldap_compare_ext_s.
However, this call is returning -3, which is not a valid LDAP return code.
This PR checks for that condition (only if not WinLDAP) and if it gets the
-3, it calls the legacy ldap_compare_s which works fine.

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>