HPCC-14678 Check/abort WU when WUSyntaxCheck times out
When ESP performs an ECL syntax check using WUSyntaxCheck, it
submits a workunit. If that workunit takes greater than the
timeout period (default 60 seconds), the existing code tries
to delete the workunit. The workunit deletion may fail if the
workunit is still in a submitted state. In this fix, the code
is added to check the workunit state. If the state is allowed,
deleteWorkUnit() will be called. If not, call secAbortWorkUnit()
then deleteWorkUnit().

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>