HPCC-14899 Create new WU only if no wuid in WUUpdate request
Since both WsWorkunits.WUUpdate and WsWorkunits.WUCreateAndUpdate
share the same request structure 'WUUpdateRequest', the correct
request for WsWorkunits.WUUpdate should be named as 'WUUpdate'.
If the request is named as 'WUUpdateRequest', WsWorkunits will
call WUCreateAndUpdate. This fix modifies the WUCreateAndUpdate
to check whether a wuid is in the 'WUUpdateRequest' or not. New
WU will be created only if no wuid in the 'WUUpdateRequest'.

A bug is fixed for adding soap request into ESP SOAP Test page.
The code similar to EspHttpBinding::generateSampleXml() is added
into EspHttpBinding::getSoapMessage() in order to find out
correct XML tag and filter SOAP request using the tag.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>