HPCC-15009 QuerySetQuery not displaying icon
Within the QueryDetails area when a query is suspended by cluster the icon appears correctly but not in the
query page. We check for a string in the Errors key but, sometimes there is no error string reported from
ClusterQueryState. Adding a check to verify both error and as long as query status is !Available
string inside of State.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>