HPCC-15084 CInterface atomics
Split eclrtl module in two to isolate all usage of boost regex as far as
possible, since (older versions of) boost have to be compiled without the
c++11 flag, but jlib can only be compiled with the c++11 flag. The isolated
code in eclregex.cpp avoids the use of jlib.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>