HPCC-15194 Log and check connection to dali
We got several reports about files being locked by ESP. Without
logging (and checking) connection ID for each lock, it may be
difficult to find out which ESP call creates a file lock and
whether the lock is a leak or not.

In this fix, an option 'LogConnection' is added to the CClient-
SDSManager. If the option is set, the SDSManager will log the
Connection ID when a connection is created to dali. I also add
the connection ID back to daliadmin dalilocks report. The
LogConnection option is configurable and default to off.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>