HPCC-14940 Limited, read-only, context information for security managers
Define IEspSecureContext interface to provide security manager plug-ins with
access to required information about the request being handled without allowing
the plug-in to affect the nature of the request. The goals of the interface
include:

 - Incur minimal overhead cost when interacting with security managers that do
   not need any information from the context. The concrete class holds a
   reference to the data it needs, and only accesses values on demand.

 - Allow each protocol to define the types of data values it supports without
   affecting the interface for existing implementations. The getProp method
   accepts any type and name values to identify a requested value.

 - Provide a uniform method to convert non-textual values to their intrinsic
   types, including checks for over- and under-flow. The interface declares and
   implements a template wrapper of getProp to support conversions. The default
   conversions are numeric, but the signature allows callers to supply custom
   converters to override or extend this behavior.

Define the TokenDeserializer class as a counterpart to the existing
TokenSerializer. This class attempts to convert a character array into a
primitive data type for which std::is_arithmetic is true.

Update the ISecManager interface to accept a pointer to IEspSecureContext for
authorization requests, and any others for which a known need exists. Update
existing security managers to accept the ISecManager changes.

Change the HTTP protocol stack to construct an HTTP-specific subclass of
IEspSecureContext, associate it with the existing IEspContext instance, and
pass it to the ISecManager instance as needed.

Signed-off-by: Tim Klemm <Tim.Klemm@lexisnexis.com>