HPCC-15199 Setting useTreeCopy may cause corrupt files on Roxie
The useTreeCopy option has never been tested in anger as far as I know (it's
been in the system for years, but no-one knows how the code works or what it
is supposed to do).

Give that the one time someone DID try it it seems to have not handled error
cases well and ended up with corrupted indexes on Roxie. the commit removes
the option and associated code altogether.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>