HPCC-15356 Possible race condition in ASyncFor can cause core in sem_post
Under very heavy load, there is a possibility that the semaphore that is
signalled to indicate a thread is complete may be unloaded part way through
the execution of sem_post, leading to a core.

Requires that the parent calling ASyncFor is in a thread which terminates during the time that the caller of
sem_post is unscheduled. Doesn't happen often as thread stacks are normally
cached and reused, but has been observed on batch roxie.

Thread::join has some (rather convoluted) code in that prevents a similar
race, so use that.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>