HPCC-14516 Logical file size reporting not correct
Within LFDetailsWidget actual size is empty. The response never returns a
value for that so I have removed it from the interface. I have also added a
string that is displayed if a file format that returns 0 is sprayed. I have
added a isCompressed flag as well to help the user with an additional clue
that the file is indeed compressed. Although, the compressed file size could
be argued is enough of a marker for that.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>
Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>